Move deserialization tests to schema
These tests tests are more appropriately placed with the schema tests.
